% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bp.basis}
\alias{bp.basis}
\title{Bernstein basis polynomials calculations}
\usage{
bp.basis(time, degree, tau = max(time))
}
\arguments{
\item{time}{a vector of times.}

\item{degree}{Bernstein polynomial degree}

\item{tau}{must be greater than times maximum value observed.}
}
\value{
A list containing matrices b and B corresponding BP basis and corresponding tau value used to compute them.
}
\description{
Bernstein basis polynomials calculations
}
