% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spbp.R
\name{bp}
\alias{bp}
\title{Bernstein Polynomials basis calculations}
\usage{
bp(time, m, tau = NULL)
}
\arguments{
\item{time}{a vector of times to events.}

\item{m}{Bernstein Polynomial degree}

\item{tau}{must be greater than times maximum value observed, this value is used for correction purposes.}

\item{data}{a data.frame with variables named in the formula.}
}
\value{
A list containing matrices b and B corresponding BP basis and corresponding tau value used to compute them.
}
\description{
Bernstein Polynomials basis calculations
}
