% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{bpph}
\alias{bpph}
\title{Bernstein Polynomials for proportional odds model (base odds estimation)}
\usage{
bpph(...)

bpph(...)
}
\arguments{
\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}

\item{formula}{a formula object, with time-to-event data on the left side of ~ and explanatory terms on the right.}

\item{m}{Bernstein Polynomial degree}

\item{tau}{must be greater than times maximum value observed, this value is used for correction purposes.}

\item{optimize}{bayes or mle inference method, by default optimize = FALSE.}

\item{data}{a data.frame with variables named in the formula.}

\item{formula}{a formula object, with time-to-event data on the left side of ~ and explanatory terms on the right.}

\item{m}{Bernstein Polynomial degree}

\item{tau}{must be greater than times maximum value observed, this value is used for correction purposes.}

\item{optimize}{bayes or mle inference method, by default optimize = FALSE.}

\item{data}{a data.frame with variables named in the formula.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`

An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bernstein Polynomials for proportional odds model (base odds estimation)

Bernstein Polynomials for proportional hazards model (base risk estimation)
}
