% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef}
\alias{coef}
\title{Generic S3 method coef}
\usage{
coef(x, ...)
}
\arguments{
\item{x}{a fitted model object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the estimated regression coefficients
}
\description{
Generic S3 method coef
}
