% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.spbp}
\alias{coef.spbp}
\title{Estimated regression coefficients}
\usage{
\method{coef}{spbp}(object, summary = c("mean", "median", "mode"), ...)
}
\arguments{
\item{object}{an object of the class spbp}

\item{summary}{posterior summary if method ="bayes" in x}

\item{...}{further arguments passed to or from other methods}
}
\value{
the estimated regression coefficients
}
\description{
Estimated regression coefficients
}
