% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{confint.spbp}
\alias{confint.spbp}
\title{Confidence intervals for the regression coefficients}
\usage{
\method{confint}{spbp}(x, level = 0.95)
}
\arguments{
\item{x}{an object of the class x}

\item{level}{the confidence level required}

\item{...}{further arguments passed to or from other methods}
}
\value{
100(1-alpha) confidence intervals for the regression coefficients
}
\description{
Confidence intervals for the regression coefficients
}
