% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{credint.spbp}
\alias{credint.spbp}
\title{Confidence intervals for the regression coefficients}
\usage{
\method{credint}{spbp}(x, prob = 0.95, type = c("HPD", "Equal-Tailed"))
}
\arguments{
\item{x}{an object of the class x}

\item{prob}{the probability level required}

\item{...}{further arguments passed to or from other methods}
}
\value{
100(1-alpha) confidence intervals for the regression coefficients
}
\description{
Confidence intervals for the regression coefficients
}
