% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{credint}
\alias{credint}
\title{Generic S3 method credint}
\usage{
credint(x, ...)
}
\arguments{
\item{x}{a fitted model object}

\item{...}{further arguments passed to parent method}
}
\value{
Generic function dispatching to \code{credint.spbp}.
}
\description{
Generic S3 method credint
}
