% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spbp.R
\name{.spbp_bayes}
\alias{.spbp_bayes}
\title{Bayesian fit for spbp (internal)}
\usage{
.spbp_bayes(standata, hessian = FALSE, verbose = FALSE, chains = 1, ...)
}
\description{
Bayesian fit for spbp (internal)
}
\keyword{internal}
