% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spbp.R
\name{.spbp_mle}
\alias{.spbp_mle}
\title{Internal: compute survival CI bands (survfit-style)}
\usage{
.spbp_mle(standata, hessian = TRUE, verbose = FALSE, ...)
}
\description{
Internal: compute survival CI bands (survfit-style)
}
\keyword{internal}
