% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.survfit_confint}
\alias{.survfit_confint}
\title{Internal: compute survival CI bands (survfit-style)}
\usage{
.survfit_confint(
  p,
  se,
  logse = TRUE,
  conf.type,
  conf.int,
  selow,
  ulimit = TRUE
)
}
\description{
Internal: compute survival CI bands (survfit-style)
}
\keyword{internal}
