% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spbp-estimates.R
\name{estimates}
\alias{estimates}
\title{Parameter estimates for fitted spbp models}
\usage{
estimates(object, ...)
}
\arguments{
\item{object}{A fitted \code{"spbp"} object.}

\item{...}{Not used.}
}
\value{
Named numeric vector of regression and Bernstein-polynomial parameters.
}
\description{
Parameter estimates for fitted spbp models
}
