% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{extract}
\alias{extract}
\title{Generic S3 method extract}
\usage{
extract(spbp, ...)
}
\arguments{
\item{spbp}{a fitted model object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
extract the MCMC chain values of a Bayesian fit.
}
\description{
Generic S3 method extract
}
