% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.spbp.R
\name{extract.spbp}
\alias{extract.spbp}
\title{Extract method for fitted spbp models}
\usage{
\method{extract}{spbp}(spbp, ...)
}
\arguments{
\item{spbp}{the result of a \code{\link[spsurv]{spbp}} fit.}

\item{...}{arguments inherent from \code{\link[rstan]{extract}}.}
}
\value{
see \code{\link[rstan]{extract}}.
}
\description{
Extract samples from a fitted \code{\link[spsurv]{spbp}} model.
}
\examples{
fit <- spbp(Surv(time, status) ~ age + sex, approach = "bayes", data = lung)
extract(fit)
}
\seealso{
\code{\link[spsurv]{spbp}}, \code{\link[spsurv]{stan_dens.spbp}}, \code{\link[spsurv]{traceplot.spbp}}
}
