% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spbp-model-comparison.R
\name{extractAIC.spbp}
\alias{extractAIC.spbp}
\title{Extract AIC from a fitted spbp model}
\usage{
\method{extractAIC}{spbp}(fit, scale = 0, k = 2, ...)
}
\arguments{
\item{fit}{A fitted \code{"spbp"} object.}

\item{scale}{Not used.}

\item{k}{Penalty per parameter (default 2).}

\item{...}{Not used.}
}
\description{
Extract AIC from a fitted spbp model
}
