% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit.R
\name{ggdiagnostics2}
\alias{ggdiagnostics2}
\title{Plot the residuals for the Cox model}
\usage{
ggdiagnostics2(coxfit, type)
}
\arguments{
\item{type}{one of "cox-snell", "martingale" or "deviance"}

\item{spbp}{A coxfit object}
}
\value{
The residuals of selected type
}
\description{
Plot the residuals for the Cox model
}
\examples{
library(survverse)
coxph.fit2 <- coxph(Surv(futime, fustat) ~ age + ecog.ps, data = ovarian)
ggcoxdiagnostics2(coxph.fit2, type = "deviance")
ggcoxdiagnostics2(coxph.fit2, type = "schoenfeld")
}
