% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{grad}
\alias{grad}
\title{Gradient for Bernstein Polynomial based regression}
\usage{
grad(t, features, coef, degree, tau = NULL, model)
}
\arguments{
\item{degree}{Bernstein Polynomial degree.}

\item{tau}{Real valued number greater than any time observed.}

\item{model}{Proportional Hazards or Proportional Odds BP based regression, default is model = "ph".}

\item{spbp}{a spbp class object.}

\item{data}{a data.frame object.}

\item{approach}{Bayesian or Maximum Likelihood estimation methods, default is approach = "bayes".}

\item{priors}{Prior settings for the Bayesian approach.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains) or `rstan::optimizing`.}
}
\value{
An object of class `stanfit` returned by `rstan::stan`.
}
\description{
Gradient for Bernstein Polynomial based regression
}
\examples{
print(fitmle)
}
\references{
Osman, M. and Ghosh, S. K. (2012), “Nonparametric regression models for right-censoreddata using Bernstein polynomials,”Computational Statistics & Data Analysis, 56, 559–573.
}
\seealso{
https://mc-stan.org/users/documentation/
}
