% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spbp-model-comparison.R
\name{logLik.spbp}
\alias{logLik.spbp}
\title{Log-likelihood for fitted spbp models}
\usage{
\method{logLik}{spbp}(object, ...)
}
\arguments{
\item{object}{A fitted \code{"spbp"} object.}

\item{...}{Not used.}
}
\description{
Log-likelihood for fitted spbp models
}
