% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{mode}
\alias{mode}
\title{Calculate the posterior mode}
\usage{
mode(ext)
}
\arguments{
\item{ext}{rstan extracted sample.}
}
\value{
A vector containing the posterior mode of each sample.
}
\description{
Calculate the posterior mode
}
