% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{model.matrix}
\alias{model.matrix}
\title{Generic S3 method model.matrix}
\usage{
model.matrix(spbp, ...)
}
\arguments{
\item{spbp}{a fitted model object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the matrix of explanatory variables of a data set.
}
\description{
Generic S3 method model.matrix
}
