% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp.basis.R
\name{plot.spbp}
\alias{plot.spbp}
\title{BP based models plot.}
\usage{
\method{plot}{spbp}(
  x,
  main,
  graph = c("baseline", "basis"),
  cumulative = F,
  frame = F,
  lwd = 3,
  ...
)
}
\arguments{
\item{x}{an object of class `spbp` result of a \code{\link[spsurv]{spbp}} fit.}

\item{main}{graph title}

\item{graph}{type of polynomial graph, default is "basis"}

\item{cumulative}{TRUE for odds and cumulative hazard}

\item{frame}{graphical parameter; default is FALSE}

\item{lwd}{graphical parameter; default is 3}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot for a fitted \code{\link[spsurv]{spbp}} model.
}
\examples{

library("spsurv")
data("veteran")

fit <- bpph(Surv(time, status) ~ karno + factor(celltype),
  data = veteran
)
plot(fit)
}
\seealso{
\code{\link[spsurv]{spbp}}.
}
