% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.bpph.bayes.R
\name{print.summary.bpph.bayes}
\alias{print.summary.bpph.bayes}
\title{Bernstein Polynomial based Regression Object}
\usage{
\method{print}{summary.bpph.bayes}(...)
}
\description{
Bernstein Polynomial based Regression Object
}
