% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.bppo.mle.R
\name{print.summary.bppo.mle}
\alias{print.summary.bppo.mle}
\title{Bernstein Polynomial based Regression Object}
\usage{
\method{print}{summary.bppo.mle}(...)
}
\arguments{
\item{...}{further arguments passed to or from other methods}
}
\description{
Bernstein Polynomial based Regression Object
}
