% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.spbp.bayes.R
\name{print.summary.spbp.bayes}
\alias{print.summary.spbp.bayes}
\title{Bernstein Polynomial Based Regression Object Summary Bayes}
\usage{
\method{print}{summary.spbp.bayes}(x, digits = 2, signif.stars = getOption("show.signif.stars"), ...)
}
\arguments{
\item{x}{a summary.spbp.bayes object}

\item{digits}{number of digits to display.}

\item{signif.stars}{see \code{\link{getOption}}}

\item{...}{further arguments passed to or from other methods}
}
\value{
none
}
\description{
Bernstein Polynomial Based Regression Object Summary Bayes
}
