% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.spbp.mle.R
\name{print.summary.spbp.mle}
\alias{print.summary.spbp.mle}
\title{Bernstein Polynomial based Regression Object}
\usage{
\method{print}{summary.spbp.mle}(x, digits = max(getOption("digits") - 4,
  3), signif.stars = getOption("show.signif.stars"), ...)
}
\description{
Bernstein Polynomial based Regression Object
}
