% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw.basis.R
\name{pw.basis}
\alias{pw.basis}
\title{Power basis polynomials calculations}
\usage{
pw.basis(degree)
}
\arguments{
\item{degree}{Bernstein polynomial degree}
}
\value{
A list containing matrices g and G corresponding BP basis and corresponding tau value used to compute them.
}
\description{
Power basis polynomials calculations
}
