% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit.R
\name{residuals.spbp}
\alias{residuals.spbp}
\title{BP based models residuals.}
\usage{
\method{residuals}{spbp}(
  object,
  type = c("martingale", "deviance", "cox-snell", "coxsnell"),
  ...
)
}
\arguments{
\item{object}{an object of class `spbp` result of a \code{\link[spsurv]{spbp}} fit.}

\item{type}{type of residuals; default is \code{"martingale"}. Also
\code{"deviance"}, \code{"cox-snell"}, and \code{"coxsnell"} (alias).}

\item{...}{arguments passed to parent method.}
}
\description{
Residuals for a fitted \code{\link[spsurv]{spbp}} model.
}
\examples{

library("spsurv")
data("veteran", package = "survival")

fit <- bpph(Surv(time, status) ~ karno + factor(celltype),
  data = veteran
)

residuals(fit)
}
\seealso{
\code{\link[spsurv:spbp]{spbp}}, \code{\link[survival:survfit]{spbp}}.
}
