% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurv.R
\name{rsurv}
\alias{rsurv}
\title{Inverse Transform Sampling To Generate Time-to-event Data From Parametric Models}
\usage{
rsurv(
  n,
  beta = c(2, -1),
  lambda = 0.5,
  k = 2,
  x = data.frame(rnorm(n), rnorm(n)),
  dist = c("weibull", "log-logistic"),
  censor = 0.25
)
}
\arguments{
\item{n}{integer; sample size}

\item{beta}{vector of regression coefficients}

\item{lambda}{event scale parameter}

\item{k}{event shape parameter}

\item{x}{matrix of independent variables (columns)}

\item{dist}{"weibull" or "log-logistic"}

\item{censor}{censoring percentage; approx. mean(status)}
}
\value{
data.frame of `ncol(x) +2` columns in which the
 survival times are the response variable denoted by `y`,
  `status` indicates failure (0 = failure) and dependent variables
  are appended to the next columns.
}
\description{
Random survival times generation for the weibull or
 log-logistic distributions with parameters `scale` and `shape`.
}
\details{
rsurv returns weibull or log-logistic(log-logistic) randomly
generated survival times. According to Collett (2003), the
accelerated failure time model encompasses a wide variety of parametric
models, including weibull and log-logistic models.
}
\examples{
rows <- 200

categorical <- rbinom(rows, size = 3, prob = .5)
x <- data.frame(numerical = rnorm(rows),
           cat0 = as.numeric(categorical == 0),
           cat1 = as.numeric(categorical == 1),
           cat2 = as.numeric(categorical == 2),
           cat3 = as.numeric(categorical == 3))

newdata <- rsurv(n = rows, beta = c(1, -2, .5, .1, 1),
  features = x, model = 'ph', dist = 'weibull')

}
\seealso{
\code{\link[spsurv]{spbp}}
}
