% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_surv.R
\name{sim_surv}
\alias{sim_surv}
\title{Inverse transform sampling from parametric models}
\usage{
sim_surv(n, beta = c(2, -1), scaleT = 2, scaleC = 3,
  x = data.frame(x1 = rnorm(n, 0), x2 = rnorm(n, 0)), shape = 2,
  model = c("ph", "po", "aft"), dist = c("weibull", "llogis"),
  censor = TRUE)
}
\arguments{
\item{n}{integer; sample size.}

\item{beta}{vector of regression coefficients.}

\item{scaleT, scaleC}{event and censoring scale parameters.}

\item{x}{matrix of features (columns).}

\item{shape}{event and censoring distribution shape.}

\item{model}{either "ph" (default) or "aft" for weibull and
"po" or "aft" for log-logistic distribution.}

\item{censor}{logical; if `TRUE`, censoring is required, that is
mean(status) > 0.}
}
\value{
data.frame of `ncol(x) +2` columns in which the
 survival times are the response variable denoted by `y`,
  `status` indicates failure (0 = failure) and the features
  are appended to the next columns.
}
\description{
Random survival times generation for the weibull or
 log-logistic distibutions with parameters `scale` and `shape`.
}
\details{
sim_surv returns weibull (log-logistic) randomly
generated survival times. According to Collett (2003), the
Accelerated Failure Time emcompasses a wide variety of parametric
models, including weibull and log-logistic models.
}
\examples{
rows <- 200

categorical <- rbinom(rows, size = 3, prob = .5)
x <- cbind(numerical = rnorm(rows),
           cat0 = as.numeric(categorical == 0),
           cat1 = as.numeric(categorical == 1),
           cat2 = as.numeric(categorical == 2),
           cat3 = as.numeric(categorical == 3))

newdata <- sim_weibull(n = rows, beta = c(1, -2, .5, .1, 1),
  x = x, model = 'ph', dist = 'weibull')

}
\references{
Collett, D. (2015). Modelling survival data in medical research. Chapman and Hall/CRC.

 Klein, J. P., & Moeschberger, M. L. (2006). Survival analysis: techniques for censored and truncated data. Springer Science & Business Media.
}
\seealso{
\code{\link[spsurv]{sim_llogis}}, \code{\link[spsurv]{spbp}}
}
