% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsurv-package.R
\docType{package}
\name{spsurv-package}
\alias{spsurv-package}
\alias{spsurv}
\title{The 'spsurv' package.}
\description{
A set of flexible routines to allow semiparametric
survival regression modeling based on Bernstein Polynomials.
\code{spsurv::spbp()} includes Proportional Hazards (BPPH), Proportional Odds (BPPO) and
Accelerated Failure Time (BPAFT) frameworks for right-censored data.
}
\references{
Demarqui, F. N., & Mayrink, V. D. (2019). A fully likelihood-based approach to model survival data with crossing survival curves. arXiv preprint arXiv:1910.02406.

Demarqui, F. N., Mayrink, V. D., & Ghosh, S. K. (2019). An Unified Semiparametric Approach to Model Lifetime Data with Crossing Survival Curves. arXiv preprint arXiv:1910.04475.

Osman, M., & Ghosh, S. K. (2012). Nonparametric regression models for right-censored data using Bernstein polynomials. Computational Statistics & Data Analysis, 56(3), 559-573.

Lorentz, G. G. (1953). Bernstein polynomials. American Mathematical Society.
}
