% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-spsurv-fit.R
\name{spsurv_fit_proportional_hazards}
\alias{spsurv_fit_proportional_hazards}
\title{Fit proportional hazards model (spsurv parsnip engine)}
\usage{
spsurv_fit_proportional_hazards(formula, data, case_weights = NULL, ...)
}
\description{
Fit proportional hazards model (spsurv parsnip engine)
}
\keyword{internal}
