% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-spsurv-fit.R
\name{spsurv_fit_proportional_odds}
\alias{spsurv_fit_proportional_odds}
\title{Fit proportional odds model (spsurv parsnip engine)}
\usage{
spsurv_fit_proportional_odds(formula, data, case_weights = NULL, ...)
}
\description{
Fit proportional odds model (spsurv parsnip engine)
}
\keyword{internal}
