% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-spsurv-fit.R
\name{spsurv_fit_survival_reg}
\alias{spsurv_fit_survival_reg}
\title{Fit AFT model (spsurv parsnip engine)}
\usage{
spsurv_fit_survival_reg(formula, data, case_weights = NULL, ...)
}
\description{
Fit AFT model (spsurv parsnip engine)
}
\keyword{internal}
