% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-spsurv-pred.R
\name{spsurv_pred_linear_pred}
\alias{spsurv_pred_linear_pred}
\title{Predict linear predictor (spsurv parsnip engine)}
\usage{
spsurv_pred_linear_pred(object, new_data, ...)
}
\description{
Predict linear predictor (spsurv parsnip engine)
}
\keyword{internal}
