% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-spsurv-pred.R
\name{spsurv_pred_survival}
\alias{spsurv_pred_survival}
\title{Predict survival probabilities (spsurv parsnip engine)}
\usage{
spsurv_pred_survival(object, new_data, eval_time, ...)
}
\description{
Predict survival probabilities (spsurv parsnip engine)
}
\keyword{internal}
