% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-spsurv-pred.R
\name{spsurv_pred_time}
\alias{spsurv_pred_time}
\title{Predict median event time (spsurv parsnip engine)}
\usage{
spsurv_pred_time(object, new_data, ...)
}
\description{
Predict median event time (spsurv parsnip engine)
}
\keyword{internal}
