% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{stan_dens.spbp}
\alias{stan_dens.spbp}
\title{Density plots method for fitted spbp models}
\usage{
\method{stan_dens}{spbp}(spbp, pars = c("beta", "gamma"), ...)
}
\arguments{
\item{spbp}{the result of a \code{\link[spsurv]{spbp}} fit.}

\item{...}{arguments inherent from \code{\link[rstan]{stan_dens}}.}
}
\value{
see \code{\link[rstan]{stan_dens}}.
}
\description{
Posterior density of samples from a fitted \code{\link[spsurv]{spbp}} model.
}
\examples{
fit <- spbp(Surv(time, status) ~ age + sex, approach = "bayes", data = lung)
stan_dens(fit)
}
\seealso{
\code{\link[spsurv]{spbp}}, \code{\link[spsurv]{traceplot.spbp}}, \code{\link[spsurv]{extract.spbp}}
}
