% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{stan_dens}
\alias{stan_dens}
\title{Generic S3 method extract}
\usage{
stan_dens(spbp, ...)
}
\arguments{
\item{spbp}{a fitted model object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the density plot of a MCMC chain.
}
\description{
Generic S3 method extract
}
