% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivor.R
\name{survivor.default}
\alias{survivor.default}
\alias{survivor.spbp}
\title{Survivor function calculations for Bernstein Polynomial based regression models}
\usage{
\method{survivor}{default}(
  time,
  arg = list(beta = NULL, gamma = NULL),
  newdata,
  model = c("ph", "po", "aft"),
  approach = c("mle", "bayes"),
  ...
)

\method{survivor}{spbp}(spbp, newdata, ...)
}
\arguments{
\item{time}{survival times}

\item{arg}{point estimates or chain values}

\item{newdata}{set of features referring to a specific population}

\item{model}{Proportional Hazards or Proportional Odds BP based regression, default is model = "ph"}

\item{approach}{Bayesian or Maximum Likelihood estimation methods, default is approach = "bayes"}

\item{...}{further arguments passed to or from other methods}

\item{spbp}{an object of the class spbp}
}
\value{
Returns the probabilities that a subject will survive beyond any given times

Returns the probabilities that a subject will survive beyond any given times
}
\description{
Default function to allow survival estimates

A method to allow survivor estimates for a spbp fit
}
\examples{

data("veteran") ## imports veteran dataset from survival package

library("spsurv")
fit <- spbp(Surv(time, status) ~ karno + factor(celltype),
data = veteran, approach =  "bayes", model = "po", chains = 1,
iter = 1000, cores = 1)

#survivor(fit)

}
