% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivor.R
\name{survivor}
\alias{survivor}
\alias{survivor.default}
\alias{survivor.spbp}
\title{Survivor function calculations for Bernstein Polynomial based regression models}
\usage{
survivor(spbp, ...)

\method{survivor}{default}(time, arg = list(beta = NULL, gamma = NULL),
  newdata, model = c("ph", "po", "aft"), approach = c("mle", "bayes"))

\method{survivor}{spbp}(spbp, newdata)
}
\arguments{
\item{spbp}{}
}
\value{
Returns the probabilities that a subject will survive beyond any given times.

Returns the probabilities that a subject will survive beyond any given times.
}
\description{
A method to ease survivor function computation.
}
\examples{
data("veteran") ## imports from survival package
library("spsurv")

fit <- spbp(Surv(time, status) ~ karno + factor(celltype),
data = veteran, approach =  "bayes", model = "po", chains = 1, iter = 1000)

survivor(fit)

}
\references{
Osman, M., & Ghosh, S. K. (2012). Nonparametric regression models for right-censored data using Bernstein polynomials. Computational Statistics & Data Analysis, 56(3), 559-573.
}
\seealso{
\code{\link[spsurv]{spbp}}, \code{\link[spsurv]{sim_surv}}
}
