% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivor.R
\name{survivor.spbp}
\alias{survivor.spbp}
\title{Survivor function calculations for an spbp object outcome}
\usage{
\method{survivor}{spbp}(spbp, time = spbp$y[, 1])
}
\arguments{
\item{n}{integer; sample size.}
}
\value{
data.frame of `ncol(x) +2` columns in which the
 survival times are the response variable denoted by `y`,
  `status` indicates failure (0 = failure) and the features
  are appended to the next columns.
}
\description{
A method to ease survivor function computation.
}
\details{
sim_surv returns weibull (log-logistic) randomly
generated survival times. According to Collett (2003), the
Accelerated Failure Time emcompasses a wide variety of parametric
models, including weibull and log-logistic models.
}
\examples{
library(spsurv)

db <- sim_surv(100, model = "aft")
fit1 <- spbp(Surv(y,status) ~ x1 + x2,  model = "aft",
          approach = "mle", data = db)

curve(survivor(time = x, spbp = fit1),
ylim = c(0,1), xlim = c(0,fit1$tau))

points(x = 2, y = survivor.spbp(time = 2, fit1), pch = 19, col = 3)

}
\references{
Osman, M., & Ghosh, S. K. (2012). Nonparametric regression models for right-censored data using Bernstein polynomials. Computational Statistics & Data Analysis, 56(3), 559-573.
}
\seealso{
\code{\link[spsurv]{sim_llogis}}, \code{\link[spsurv]{spbp}}
}
