% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpph.R
\name{terms.inner}
\alias{terms.inner}
\title{A function taken from the survival library - it is not exported from there hence a local copy}
\usage{
\method{terms}{inner}(x)
}
\arguments{
\item{x}{an R object}
}
\value{
a character vector
}
\description{
A function taken from the survival library - it is not exported from there hence a local copy
}
\keyword{internal}
