% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{traceplot}
\alias{traceplot}
\title{Generic S3 method traceplot}
\usage{
traceplot(spbp, ...)
}
\arguments{
\item{spbp}{a fitted model object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the traceplot of a MCMC chain.
}
\description{
Generic S3 method traceplot
}
