% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceplot.spbp.R
\name{traceplot.spbp}
\alias{traceplot.spbp}
\title{Traceplot method for fitted spbp models}
\usage{
traceplot.spbp(spbp, ...)
}
\arguments{
\item{spbp}{the result of a \code{\link[spsurv]{spbp}} fit.}

\item{...}{arguments inherent from \code{\link[rstan]{traceplot}}.}
}
\value{
see \code{\link[rstan]{traceplot}}.
}
\description{
Traceplot samples from a fitted \code{\link[spsurv]{spbp}} model.
}
\examples{
fit <- spbp(Surv(time, status) ~ age + sex, approach = "bayes", data = lung)
traceplot(fit)
}
\seealso{
\code{\link[spsurv]{spbp}}, \code{\link[spsurv]{stan_dens.spbp}}, \code{\link[spsurv]{extract.spbp}}
}
