% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpph.R
\name{bpph}
\alias{bpph}
\title{Bernstein Polynomials based Proportional Hazards}
\usage{
bpph(formula, data, ...)
}
\arguments{
\item{formula}{a Surv object with time to event, status and explanatory terms.}

\item{data}{a data.frame object.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains) or `rstan::optimizing`.}

\item{degree}{Bernstein Polynomial degree.}

\item{tau}{Real valued number greater than any time observed.}

\item{approach}{Bayesian or Maximum Likelihood estimation methods, default is approach = "bayes".}

\item{model}{Proportional Hazards or Proportional Odds BP based regression, default is model = "ph".}

\item{priors}{Prior settings for the Bayesian approach.}
}
\value{
An object of class `stanfit` returned by `rstan::stan`.
}
\description{
Fits Bernstein Polynomial based Proportional Odds model to lung cancer data.
}
\examples{

data("veteran") ## imports from survival package
library("spsurv")

fit <- spbp(Surv(time, status) ~ karno + factor(celltype),
data = veteran, approach =  "bayes", model = "po", chains = 1, iter = 1000)

print(fit)
summary(fit)
}
\references{
Osman, M. and Ghosh, S. K. (2012), “Nonparametric regression models for right-censoreddata using Bernstein polynomials,”Computational Statistics & Data Analysis, 56, 559–573.
}
\seealso{
\url{https://mc-stan.org/users/documentation/}
}
