% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpph.R
\name{bpph}
\alias{bpph}
\title{Bernstein Polynomials base risk estimation in Proportional hazards model}
\usage{
bpph(formula, m = ceiling(sqrt(nrow(data))), tau = NA, data,
  approach = c("bayesian", "frequentist"), prior = NULL, chains = 1,
  verbose = FALSE, ...)
}
\arguments{
\item{formula}{a formula object, with time-to-event data on the left side of ~ and explanatory terms on the right.}

\item{m}{Bernstein Polynomial degree}

\item{tau}{must be greater than times maximum value observed, this value is used for correction purposes.}

\item{data}{a data.frame with variables named in the formula.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}

\item{optimize}{Bayesian or frequentist inference method, by default optimize = FALSE.}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Bernstein Polynomials base risk estimation in Proportional hazards model
}
