% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpfits.R
\name{bpph}
\alias{bpph}
\title{Bernstein polynomial based proportional hazards model}
\usage{
bpph(formula, degree, data, approach = "mle", ...)
}
\arguments{
\item{formula}{a Surv object with time to event observations, right censoring status and explanatory terms.}

\item{degree}{Bernstein polynomial degree.}

\item{data}{a data.frame object.}

\item{approach}{Bayesian or maximum likelihood estimation methods, default is approach = "mle".}

\item{priors}{Prior choice for the Bayesian analysis, see \code{\link[spsurv]{spbp}}.}
}
\value{
An object of class `spbp`.
}
\description{
Fits the BPPH model to time-to-event data.
}
\examples{

library("KMsurv")
data("larynx")

library("spsurv")
fit <- bpph(Surv(time, status) ~ age + factor(stage),
data = larynx)

summary(fit)

}
\references{
Osman, M. and Ghosh, S. K. (2012), “Nonparametric regression models for right-censoreddata using Bernstein polynomials,”Computational Statistics & Data Analysis, 56, 559–573.
}
\seealso{
\code{\link[spsurv]{spbp}}, \code{\link[spsurv]{bppo}} and \code{\link[spsurv]{bpaft}} for other BP based models.
}
