% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coef.spbp}
\alias{coef.spbp}
\title{Estimated regression coefficients}
\usage{
\method{coef}{spbp}(spbp, ...)
}
\arguments{
\item{spbp}{an object of the class spbp}

\item{...}{further arguments passed to or from other methods}
}
\value{
the estimated regression coefficients
}
\description{
Estimated regression coefficients
}
