% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{confint.spbp}
\alias{confint.spbp}
\title{Confidence intervals for the regression coefficients}
\usage{
\method{confint}{spbp}(object, parm = names(coef(object)), level = 0.95, ...)
}
\arguments{
\item{object}{a fitted model object.}

\item{parm}{a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{the confidence level required.}

\item{...}{further arguments passed to parent method}
}
\value{
100(1-alpha) confidence intervals for the regression coefficients
}
\description{
Confidence intervals for the regression coefficients
}
