% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{credint}
\alias{credint}
\title{Generic S3 method credint}
\usage{
credint(x, ...)
}
\arguments{
\item{x}{a fitted model object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the estimated regression coefficients
}
\description{
Generic S3 method credint
}
