% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spbp.R
\name{.spbp_bayes}
\alias{.spbp_bayes}
\title{Internal: compute survival CI bands (survfit-style)}
\usage{
.spbp_bayes(standata, hessian = FALSE, verbose = FALSE, chains = 1, ...)
}
\description{
Internal: compute survival CI bands (survfit-style)
}
\keyword{internal}
