% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{model.matrix.spbp}
\alias{model.matrix.spbp}
\title{Model.matrix method for fitted spbp models}
\usage{
\method{model.matrix}{spbp}(
  object,
  data = eval(object$call$data, envir = parent.frame()),
  ...
)
}
\arguments{
\item{object}{an object of class `spbp`, see \code{\link[spsurv]{spbp}}.}

\item{data}{data.frame object.}

\item{...}{arguments inherent from \code{\link[stats]{model.matrix}}.}
}
\value{
The explanatory variables matrix.
}
\description{
Model.matrix of a fitted \code{\link[spsurv]{spbp}} model.
}
\examples{


library("spsurv")
data("veteran")

fit <- bpph(Surv(time, status) ~ karno + factor(celltype),
data = veteran)

model.matrix(fit)

}
\seealso{
\code{\link[spsurv]{spbp}}, \code{\link[stats]{model.matrix}}
}
