% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.R
\name{model.matrix.spbp}
\alias{model.matrix.spbp}
\title{Model.matrix method for fitted spbp models}
\usage{
\method{model.matrix}{spbp}(object, ...)
}
\arguments{
\item{object}{an object of class `spbp`, see \code{\link[spsurv]{spbp}}.}

\item{...}{arguments passed to parent method.}
}
\value{
The model matrix.
}
\description{
Model.matrix of a fitted \code{\link[spsurv]{spbp}} model.
}
\examples{

library("spsurv")
data("veteran")

fit <- bpph(Surv(time, status) ~ karno + factor(celltype),
  data = veteran
)

model.matrix(fit)
}
\seealso{
\code{\link[spsurv]{spbp}}, \code{\link[stats]{model.matrix}}
}
