% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.spbp.R
\name{model.matrix.spbp}
\alias{model.matrix.spbp}
\title{Model.matrix method for fitted spbp models}
\usage{
\method{model.matrix}{spbp}(spbp, data = eval(spbp$call$data, envir =
  parent.frame()))
}
\arguments{
\item{object}{the result of a \code{\link[spsurv]{spbp}} fit.}

\item{...}{arguments inherent from \code{\link[stats]{model.matrix}}.}
}
\value{
The model matrix for the fit.
}
\description{
Model.matrix of a fitted \code{\link[spsurv]{spbp}} model.
}
\examples{
fit <- spbp(Surv(time, status) ~ age + sex, approach = "bayes", data = lung)
model.matrix(fit)
}
\seealso{
\code{\link[spsurv]{spbp}}, \code{\link[stats]{model.matrix}}
}
