% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.spbp.R
\name{print.spbp}
\alias{print.spbp}
\title{Bernstein Polynomial Based Regression Object Print}
\usage{
\method{print}{spbp}(
  x,
  digits = max(getOption("digits") - 4, 3),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{an object of class spbp}

\item{digits}{number of digits to display}

\item{signif.stars}{see \code{\link{getOption}}}

\item{...}{further arguments passed to or from other methods}
}
\value{
none
}
\description{
Bernstein Polynomial Based Regression Object Print
}
