% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.spbp.R
\name{print.spbp}
\alias{print.spbp}
\title{Bernstein Polynomial based Regression Object}
\usage{
\method{print}{spbp}(
  spbp,
  digits = max(getOption("digits") - 4, 3),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\description{
Bernstein Polynomial based Regression Object
}
