% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.bpaft.bayes.R
\name{print.summary.bpaft.bayes}
\alias{print.summary.bpaft.bayes}
\title{Bernstein Polynomial Based Regression Object Summary BPAFT Bayes}
\usage{
\method{print}{summary.bpaft.bayes}(...)
}
\arguments{
\item{...}{further arguments passed to or from other methods}
}
\value{
none
}
\description{
Bernstein Polynomial Based Regression Object Summary BPAFT Bayes
}
