% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.bpaft.mle.R
\name{print.summary.bpaft.mle}
\alias{print.summary.bpaft.mle}
\title{Bernstein Polynomial Based Regression Object Summary BPAFT MLE}
\usage{
\method{print}{summary.bpaft.mle}(...)
}
\arguments{
\item{...}{further arguments passed to or from other methods}
}
\value{
none
}
\description{
Bernstein Polynomial Based Regression Object Summary BPAFT MLE
}
