% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.bpph.bayes.R
\name{print.summary.bpph.bayes}
\alias{print.summary.bpph.bayes}
\title{Bernstein Polynomial Based Regression Object Summary BPPH Bayes}
\usage{
\method{print}{summary.bpph.bayes}(...)
}
\arguments{
\item{...}{further arguments passed to or from other methods}
}
\value{
none
}
\description{
Bernstein Polynomial Based Regression Object Summary BPPH Bayes
}
