% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.bpph.mle.R
\name{print.summary.bpph.mle}
\alias{print.summary.bpph.mle}
\title{Bernstein Polynomial Based Regression Object Summary BPPH MLE}
\usage{
\method{print}{summary.bpph.mle}(...)
}
\arguments{
\item{...}{further arguments passed to or from other methods}
}
\value{
none
}
\description{
Bernstein Polynomial Based Regression Object Summary BPPH MLE
}
