% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.bppo.bayes.R
\name{print.summary.bppo.bayes}
\alias{print.summary.bppo.bayes}
\title{Bernstein Polynomial Based Regression Object Summary BPPO Bayes}
\usage{
\method{print}{summary.bppo.bayes}(...)
}
\arguments{
\item{...}{further arguments passed to or from other methods}
}
\value{
none
}
\description{
Bernstein Polynomial Based Regression Object Summary BPPO Bayes
}
