% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.spbp.bayes.R
\name{print.summary.spbp.bayes}
\alias{print.summary.spbp.bayes}
\title{Bernstein Polynomial based Regression Object}
\usage{
\method{print}{summary.spbp.bayes}(x, digits = max(getOption("digits") -
  3, 3))
}
\description{
Bernstein Polynomial based Regression Object
}
