% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.spbp.mle.R
\name{print.summary.spbp.mle}
\alias{print.summary.spbp.mle}
\title{Bernstein Polynomial Based Regression Object Summary MLE}
\usage{
\method{print}{summary.spbp.mle}(
  x,
  digits = max(getOption("digits") - 4, 3),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{a summary.spbp.mle object}

\item{digits}{number of digits to display.}

\item{signif.stars}{see \code{\link{getOption}}}

\item{...}{further arguments passed to or from other methods}
}
\value{
none
}
\description{
Bernstein Polynomial Based Regression Object Summary MLE
}
