% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivor.R
\name{residuals.spbp}
\alias{residuals.spbp}
\title{BP based models residuals.}
\usage{
\method{residuals}{spbp}(spbp, type = c("cox-snell"))
}
\arguments{
\item{spbp}{an object of class `spbp` result of a \code{\link[spsurv]{spbp}} fit.}

\item{type}{type of residuals, default is "cox-snell".}
}
\description{
Residuals for a fitted \code{\link[spsurv]{spbp}} model.
}
\examples{

data("veteran") ## imports veteran dataset from survival package

library("spsurv")

fit <- spbp(Surv(time, status) ~ karno + factor(celltype),
data = veteran, approach =  "bayes", model = "po", chains = 1, iter = 1000)

residuals(fit)

}
\seealso{
\code{\link[spsurv]{spbp}}.
}
