% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spbp.R
\name{spbp.default}
\alias{spbp.default}
\title{spbp: The BP based semiparametric survival analysis function}
\usage{
\method{spbp}{default}(
  formula,
  degree,
  data,
  approach = c("mle", "bayes"),
  model = c("ph", "po", "aft"),
  priors = list(beta = c("normal(0,4)"), gamma = "lognormal(0,10)"),
  scale = TRUE,
  ...
)
}
\value{
An object of class \code{spbp}
}
\description{
spbp: The BP based semiparametric survival analysis function
}
