% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spbp.R
\name{spbp}
\alias{spbp}
\title{Semiparametric survival analysis using Bernstein Polynomials}
\usage{
spbp(formula, degree = NULL, tau = NULL, data, approach = c("bayes",
  "mle"), model = c("ph", "po"), priors = list(shape_gamma = 0.01,
  rate_gamma = 0.01, mean_beta = 0, sd_beta = 10), verbose = FALSE, ...)
}
\arguments{
\item{formula}{a formula object, with time-to-event data on the left side of ~ and explanatory terms on the right.}

\item{degree}{Bernstein Polynomial degree.}

\item{tau}{must be greater than times maximum value observed, used for correction purposes.}

\item{data}{a data.frame with variables named in the formula.}

\item{approach}{bayes or mle methods, default is approach = "bayes".}

\item{model}{proportional hazards or porportional odds base, default is model = ph}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains), see https://mc-stan.org/users/documentation/.}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Semiparametric survival analysis using Bernstein Polynomials
}
