% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spbp.R
\name{spbp}
\alias{spbp}
\title{spbp: The BP based survival analysis function}
\usage{
spbp(formula, ...)
}
\arguments{
\item{formula}{a Surv object with time to event, status and explanatory terms.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains) or `rstan::optimizing`.}
}
\description{
Semiparametric survival analysis using Bernstein Polynomial
}
\details{
Fits Bernstein Polynomial based Proportional regression to survival data.
}
\examples{

library("spsurv")
data("veteran") ## imports from survival package

fit_mle <- spbp(Surv(time, status) ~ karno + factor(celltype),
 data = veteran, model = "po")
summary(fit_mle)

fit_bayes <- spbp(Surv(time, status) ~ karno + factor(celltype),
                  data = veteran, model = "po", approach = "bayes",
                   cores = 1, iter = 1000,
                    priors = list(beta = c("normal(0,4)"),
                     gamma = "lognormal(0,10)"))
summary(fit_bayes)

}
\references{
Osman, M. and Ghosh, S. K. (2012), “Nonparametric regression models for right-censoreddata using Bernstein polynomials,”Computational Statistics & Data Analysis, 56, 559–573.
}
\seealso{
\url{https://mc-stan.org/users/documentation/}
}
