% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spbp.R
\name{spbp}
\alias{spbp}
\title{Semiparametric survival analysis using Bernstein Polynomials}
\usage{
spbp(formula, degree = ceiling(sqrt(nrow(data))), tau = max(time),
  data, approach = c("mle", "bayes"), model = c("ph", "po", "aft"),
  priors = list(hyper_gamma = 0.1, mean_beta = 0, sd_beta = sqrt(10)),
  hessian = TRUE, verbose = FALSE, init = 0, algorithm = "LBFGS",
  chains = 1, ...)
}
\arguments{
\item{formula}{a Surv object with time to event, status and explanatory terms.}

\item{degree}{Bernstein Polynomial degree.}

\item{tau}{Real valued number greater than any time observed.}

\item{data}{a data.frame object.}

\item{approach}{Bayesian or Maximum Likelihood estimation methods, default is approach = "bayes".}

\item{model}{Proportional Hazards or Proportional Odds BP based regression, default is model = "ph".}

\item{priors}{Prior settings for the Bayesian approach.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains) or `rstan::optimizing`.}
}
\value{
An object of class `stanfit` returned by `rstan::stan`.
}
\description{
Semiparametric survival analysis using Bernstein Polynomials
}
\examples{
library(KMsurv)
data("larynx")

#Fit a Proportional Odds BP based model to the veteran cancer data set
library ("survival")
data("veteran")

library("spsurv")

fitmle <− spbp(Surv(time, delta) ~ karno + factor(celltype),
data = veteran, approach =  "mle", model = "po")

print(fitmle)
}
\references{
Osman, M. and Ghosh, S. K. (2012), “Nonparametric regression models for right-censoreddata using Bernstein polynomials,”Computational Statistics & Data Analysis, 56, 559–573.
}
\seealso{
\url{https://mc-stan.org/users/documentation/}
}
