% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spbp.R
\name{spbp}
\alias{spbp}
\alias{spbp.default}
\title{spbp: The BP based semiparametric survival analysis function}
\usage{
spbp(formula, ...)

\method{spbp}{default}(formula, degree, data, approach = c("mle",
  "bayes"), model = c("ph", "po", "aft"), priors = list(beta =
  c("normal(0,4)"), gamma = "lognormal(0,4)"), scale = TRUE, ...)
}
\arguments{
\item{formula}{a Surv object with time to event, status and explanatory terms.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains) or `rstan::optimizing`.}

\item{degree}{Bernstein Polynomial degree.}

\item{data}{a data.frame object.}

\item{approach}{Bayesian or Maximum Likelihood estimation methods, default is approach = "bayes".}

\item{model}{Proportional Hazards or Proportional Odds BP based regression, default is model = "ph".}

\item{priors}{Prior settings for the Bayesian approach.}

\item{tau}{Real valued number greater than any time observed.}
}
\value{
An object of class \code{spbp}
}
\description{
Semiparametric survival analysis using Bernstein Polynomials
}
\details{
Fits Bernstein Polynomial based Proportional regression to survival data.
}
\examples{

data("veteran") ## imports from survival package
library("spsurv")

fit <- spbp(Surv(time, status) ~ karno + factor(celltype),
data = veteran, approach =  "bayes", model = "po", chains = 1, iter = 1000)

print(fit)
summary(fit)
}
\references{
Osman, M. and Ghosh, S. K. (2012), “Nonparametric regression models for right-censoreddata using Bernstein polynomials,”Computational Statistics & Data Analysis, 56, 559–573.
}
\seealso{
\url{https://mc-stan.org/users/documentation/}
}
