% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsurv-package.R
\docType{package}
\name{spsurv-package}
\alias{spsurv-package}
\alias{spsurv}
\title{The 'spsurv' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2019). RStan: the R interface to Stan. R package version 2.19.1. https://mc-stan.org
}
