% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsurv-package.R
\docType{package}
\name{spsurv-package}
\alias{spsurv-package}
\alias{spsurv}
\title{The 'spsurv' package.}
\description{
Semiparametric Survival Analysis
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
