% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.spbp.R
\name{summary.spbp}
\alias{summary.spbp}
\title{Bernstein Polynomial Based Regression Object Summary}
\usage{
\method{summary}{spbp}(object, interval = 0.95, ...)
}
\arguments{
\item{object}{an object of class spbp}

\item{interval}{interval coverage (confidence or credibility)}

\item{...}{further arguments passed to or from other methods}
}
\value{
An object of class analogous to for e.g. 'summary.bppo.bayes'.
}
\description{
Bernstein Polynomial Based Regression Object Summary
}
