% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.spbp.R
\name{summary.spbp}
\alias{summary.spbp}
\title{Bernstein Polynomial based Regression Object}
\usage{
\method{summary}{spbp}(spbp, interval = 0.95, ...)
}
\description{
Bernstein Polynomial based Regression Object
}
