% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit.R
\name{survfit.spbp}
\alias{survfit.spbp}
\title{BP based models survfit.}
\usage{
\method{survfit}{spbp}(
  x,
  newdata,
  times,
  se.fit = TRUE,
  interval = 0.95,
  type = c("log", "log-log", "plain"),
  ...
)
}
\arguments{
\item{x}{an object of class `spbp` result of a \code{\link[spsurv]{spbp}} fit.}

\item{...}{further arguments passed to survfit.}
}
\description{
Survival for a fitted \code{\link[spsurv]{spbp}} model.
}
\examples{

library("spsurv")
data("veteran")

fit <- bpph(Surv(time, status) ~ karno + factor(celltype),
  data = veteran
)

survfit(fit)
}
\seealso{
\code{\link[spsurv]{spbp}}, \code{\link[survfit]{spbp}}.
}
