% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivor.R
\name{survivor}
\alias{survivor}
\title{Generic S3 method vcov}
\usage{
survivor(spbp, ...)
}
\arguments{
\item{spbp}{an object of class spbp}

\item{...}{further arguments passed to or from other methods}
}
\value{
estimates survival for each dataset individual (line).
Spbp Object Observed Survival
}
\description{
Generic S3 method vcov
}
\seealso{
\code{\link[spsurv]{spbp}}, \code{\link[spsurv]{itsamp}}
}
