% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivor.R
\name{survivor}
\alias{survivor}
\title{Survivor function calculations for Bernstein Polynomial based regression models}
\usage{
survivor(spbp, ...)
}
\arguments{
\item{spbp}{#' @examples
data("veteran") ## imports veteran dataset from survival package

library("spsurv")

fit <- spbp(Surv(time, status) ~ karno + factor(celltype),
data = veteran, approach =  "bayes", model = "po", chains = 1, iter = 1000)

survivor(fit)}
}
\description{
A method to allow survivor function computation.
}
\references{
Osman, M., & Ghosh, S. K. (2012). Nonparametric regression models for right-censored data using Bernstein polynomials. Computational Statistics & Data Analysis, 56(3), 559-573.
}
\seealso{
\code{\link[spsurv]{spbp}}, \code{\link[spsurv]{sim_surv}}
}
