% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vcov.spbp}
\alias{vcov.spbp}
\title{Covariance of the regression coefficients}
\usage{
\method{vcov}{spbp}(spbp, ...)
}
\arguments{
\item{spbp}{an object of the class spbp}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the variance-covariance matrix associated with the regression coefficients.
}
\description{
Covariance of the regression coefficients
}
