% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.spbp}
\alias{vcov.spbp}
\title{Covariance of the regression coefficients}
\usage{
\method{vcov}{spbp}(object, bp.param = FALSE, ...)
}
\arguments{
\item{object}{an object of the class spbp}

\item{bp.param}{return Bernstein Polynomial variance.}

\item{...}{arguments passed to parent method.}
}
\value{
the variance-covariance matrix associated with the regression coefficients.
}
\description{
Covariance of the regression coefficients
}
